requirements.txt
pymongo
playwright
requests
python-dotenv
